diary_backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6

SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0

pydantic==2.9.2
pydantic-settings==2.6.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

redis==5.2.0
boto3==1.35.61

httpx==0.27.2

pytest==8.3.3
pytest-asyncio==0.24.0
email-validator==2.2.0
